Added SirenDateExtension. Improved app safety.
